moved cover functions from nodeservice to coverService
