+++ backend/src/services/NodeService.d.ts
@@ -36,8 +36,6 @@ export function getIfPageRead(client: Client, page_id: string, user_id: string):
 
 export function getCoverById(client: Client, cover_id: string): void;
 
-export function getCoverByPageId(client: Client, page_id: string): void;
-
 export function getPageById(client: Client, page_id: string): void;
 
 export function getCovers(client: Client, limit: string): void;
@@ -70,7 +68,3 @@ export function getContinueReadingByUserId(client: Client, user_id: string): voi
 export function togglePin(client: Client, userId: string, pageId: string): void;
 
 export function getPinByUserByPage(client: Client, userId: string, pageId: string): void;
-
-export function getContributionsByUserId(client: Client, user_id: string);
-
-export function addCoverIdsToPages(client: Client): void;
