+++ backend/src/services/cover/CoverService.d.ts
@@ -1,0 +1,3 @@
+import { Client } from "pg";
+
+export function getCoverByPageId(client: Client, page_id: string): void;
